cleaning up matching algo directory
