+++ src/matching-algo/__tests__/original-match-algo.test.ts
@@ -1,4 +1,4 @@
-import { makeMatches } from './matchAlgo';
+import { makeMatches } from '../original-match-algo';
 
 let fallBackUser;
