+++ src/matching-algo/original-match-algo.ts
@@ -1,4 +1,8 @@
-import { IUser, IpastMatchObj } from './matching-algo';
+/** TODO:
+ *
+ *  - move stable marriage match
+ */
+
 import { deepClone } from '../utils/clone';
 import {
   filterForUniqueMatches,
@@ -6,80 +10,6 @@ import {
   sortByOldestMatch,
   findAndRemoveUserFromPool
 } from './matchingHelperFn';
-
-export type Acceptor = {
-  self: IUser;
-  topSuitor: IUser;
-  priority: IUser[];
-};
-
-export type Suitor = {
-  self: IUser;
-  priority: IUser[];
-  currentlyAccepted: boolean;
-};
-
-export function trimAfterRank<T>(priorityList: T[], rank: number): T[] {
-  priorityList.splice(rank + 1, priorityList.length - (rank + 1));
-
-  return priorityList;
-}
-
-export function makeStableMarriageMatches(
-  suitors: Map<string, Suitor>,
-  acceptors: Map<string, Acceptor>
-): Array<[IUser, IUser]> {
-  if (suitors.size !== acceptors.size) {
-    throw new Error('suitors and acceptor arrays not equal length');
-  }
-
-  const acceptedAcceptors = [];
-  while (acceptedAcceptors.length !== acceptors.size) {
-    for (const s of suitors.values()) {
-      if (!s.currentlyAccepted && s.priority.length !== 0) {
-        const potentialAcceptor = acceptors.get(s.priority[0].email);
-        const rank = potentialAcceptor.priority.indexOf(s.self);
-        // case 1: no current proposals --> accept
-        if (!potentialAcceptor.topSuitor) {
-          potentialAcceptor.topSuitor = s.self;
-          acceptedAcceptors.push(potentialAcceptor);
-          trimAfterRank(potentialAcceptor.priority, rank);
-          s.currentlyAccepted = true;
-        } else {
-          const currentTopSuitorRank = potentialAcceptor.priority.indexOf(
-            potentialAcceptor.topSuitor
-          );
-          // case 2: has proposal, but this suitor is worse or not in priority list --> so keep current proposal
-          if (currentTopSuitorRank < rank || rank === -1) {
-            s.priority.shift();
-          } else {
-            // case 3: has proposal, but this suitor is better --> set current proposal to this suitor, & get reference to previous suitor & set currentAccepted
-            const rejectedSuitor = suitors.get(
-              potentialAcceptor.topSuitor.email
-            );
-            rejectedSuitor.currentlyAccepted = false;
-
-            potentialAcceptor.topSuitor = s.self;
-            trimAfterRank(potentialAcceptor.priority, rank);
-            s.currentlyAccepted = true;
-          }
-        }
-      }
-    }
-  }
-
-  // make matches from acceptors accepted suitors
-  const matches: Array<[IUser, IUser]> = acceptedAcceptors.map(makeMatch);
-  //   const matches = acceptedAcceptors.map((a: Acceptor) => {
-  //     return [a.self, a.topSuitor];
-  //   });
-
-  return matches;
-}
-
-function makeMatch(a: Acceptor): [IUser, IUser] {
-  return [a.self, a.topSuitor];
-}
 
 export function makeMatches(
   usersToMatch: IUser[],
